BookingTimeForm (fix race cond): FormSpy calls line-items twice. If booking start time is moved backwards, there are 2 calls: - new start time & old end time - new start time & updated end time
If the former is returned last, a wrong hour count is displayed on the listing page.
However, this hour-counting bug does not move forward to checkout page since only
the line-items are async calls on the listing page. Checkout page has gotten updated start and end times.

This commit fixes the bug by completely removing FormSpy.