this was implemented incorrectly
not sure why affiliateProgram was set to disabled by your team. Users haven't been able to unselect it since deploying. Also i commented out the the original logic for disabling promoCode but you can see the intention was only to disable it if it didn't have a value... its now disabled always. that one's less concerning but incorrect