Fix SelectSingleFilter initial values handling for non-nested enums
SelectSingleFilter's FilterPopup cancel has an issue because the filter sets level numbers to attribute names, and those level numbers aren't present in initialValues. Fix the issue by passing isNestedEnum to the pickInitialValuesForFieldSelectTree function and setting level number 1 to the initial value keys when isNestedEnum is false. This addresses the same behavior for all non-nested filters using the SelectSingleFilter component.