cache.js: add a simple createLRUCache function
Note: this is not battle-tested with a big cache turnover.
If you are looking for a performant battle tested LRU cache,
you should check lru-cache and other libraries dedicated for caching.

The initial focus for this is caching assets related to Page content,
which on individual marketplace might easily fit in memory.